updating numbering system for sequence
made it so when person selects that it's a part of the course, it defaults to seq 1. If they uncheck it, it defaults to seq 0.

seq 0 by default since is_seq is false by default

removed check in the sequence rules that ensures seq must start at 1 now, 0 is allowed